Move edit session to seperate module
